move <Game> to component
